showcases all tokens, update gitignore
