+++ site/.gitignore
@@ -11,4 +11,4 @@ next-env.d.ts
 /snapshots/
 /public/sitemap.xml
 /src/props/*.json
-/src/components/css-display/cssVariables.json
+/src/components/css-display/json/
