--- site/docs/foundations/characteristics/accordionView.mdx
@@ -1,8 +1,0 @@
----
-title: Characteristic types
-layout: DetailTechnical
-sidebar:
-  label: Characteristic types
----
-
-<AccordionView />
