+++ site/docs/foundations/tokens/deprecated.mdx
@@ -1,0 +1,9 @@
+---
+title: Deprecated tokens
+layout: DetailTechnical
+sidebar:
+  label: Deprecated tokens
+  priority: 1
+---
+
+<CSSVariables value="deprecated" />
