+++ site/docs/foundations/tokens/foundations.mdx
@@ -1,0 +1,8 @@
+---
+title: Foundation tokens
+layout: DetailTechnical
+sidebar:
+  label: Foundation tokens
+---
+
+<CSSVariables value="foundations" />
