+++ site/docs/foundations/tokens/index.mdx
@@ -1,0 +1,9 @@
+---
+title: Tokens
+layout: DetailTechnical
+sidebar:
+  label: Tokens
+  priority: 1
+---
+
+Here is the complete set of Salt tokens.
