+++ site/docs/foundations/tokens/palette.mdx
@@ -1,0 +1,8 @@
+---
+title: Palette tokens
+layout: DetailTechnical
+sidebar:
+  label: Palette tokens
+---
+
+<CSSVariables value="palette" />
