--- site/src/components/css-display/descriptions.ts
@@ -1,76 +1,0 @@
-export default [
-  {
-    key: "accent",
-    value:
-      "Components which can or should be emphasized against others in close proximity.",
-  },
-  {
-    key: "actionable",
-    value:
-      "Components with the ability to action something, action is performed immediately (the component doesn’t stay selected).    ",
-  },
-  {
-    key: "container",
-    value:
-      "Group of components used to contain and separate different types of content and allow hierarchical organization.    ",
-  },
-  {
-    key: "draggable",
-    value:
-      "Components that can be grabbed and moved to a target area (related to Target).    ",
-  },
-  {
-    key: "editable",
-    value: "Components that hold mutable data and allow for data entry.    ",
-  },
-  {
-    key: "focusable",
-    value:
-      "Components which can be focused using the mouse or keyboard.The following variables can be used on their own, but make up parts of the more commonly used borders and outlines below.",
-  },
-  {
-    key: "navigable",
-    value:
-      "Components that allow a user to navigate across UI sections or data sets and change view context.    ",
-  },
-  {
-    key: "overlayable",
-    value:
-      "Components that may form a hierarchy and stack upon or beneath other elements of the UI.    ",
-  },
-  {
-    key: "selectable",
-    value:
-      "Group of components that allow single or multiple selection, Selection can trigger actions.    ",
-  },
-  {
-    key: "separable",
-    value:
-      " Group of attributes to separate information/elements. Connected to Container. Level 1 - used when subtle division is needed/to divide similar items e.g. options within a list, individual comments. Level 2 - used to divide related areas e.g. heading and body text. Level 3 - used to separate areas of content that are not closely related e.g. app header from main content",
-  },
-  {
-    key: "status",
-    value:
-      " Components which have attributes that denote status and the severity of that status.    ",
-  },
-  {
-    key: "taggable",
-    value:
-      "Components allowing for individual items or data to be organized.    ",
-  },
-  {
-    key: "target",
-    value:
-      "Components which indicate a target area where a draggable item can be dropped (related to Draggable).    ",
-  },
-  {
-    key: "text",
-    value:
-      " All text examples are shown using the default fontWeight token.    ",
-  },
-  {
-    key: "track",
-    value:
-      "Components which have the ability to visually indicate progress, scale, or range along a defined track.    ",
-  },
-];
