--- site/src/components/css-display/index.ts
@@ -1,1 +1,0 @@
-export * from "./AccordionView";
